sccadmission: move scc authz check into its own file
